refactor: 열거타입 클래스 파일로 추출
