+++ src/main/kotlin/kr/kro/dearmoment/common/validation/EnumValue.kt
@@ -1,0 +1,16 @@
+package kr.kro.dearmoment.common.validation
+
+import jakarta.validation.Constraint
+import jakarta.validation.Payload
+import kotlin.reflect.KClass
+
+@Target(AnnotationTarget.FIELD)
+@Retention(AnnotationRetention.RUNTIME)
+@Constraint(validatedBy = [EnumValidator::class, EnumListValidator::class])
+annotation class EnumValue(
+    val enumClass: KClass<out Enum<*>>,
+    val message: String = "",
+    val groups: Array<KClass<*>> = [],
+    val payload: Array<KClass<out Payload>> = [],
+    val ignoreCase: Boolean = false,
+)
