+++ src/main/kotlin/kr/kro/dearmoment/product/application/usecase/delete/DeleteProductOptionUseCaseImpl.kt
@@ -4,7 +4,7 @@ import kr.kro.dearmoment.common.exception.CustomException
 import kr.kro.dearmoment.common.exception.ErrorCode
 import kr.kro.dearmoment.product.application.port.out.GetProductOptionPort
 import kr.kro.dearmoment.product.application.port.out.ProductOptionPersistencePort
-import kr.kro.dearmoment.product.domain.model.ProductOption
+import kr.kro.dearmoment.product.domain.model.option.ProductOption
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
