+++ src/main/kotlin/kr/kro/dearmoment/product/domain/model/option/ProductOptionUpdateResult.kt
@@ -1,4 +1,4 @@
-package kr.kro.dearmoment.product.domain.model
+package kr.kro.dearmoment.product.domain.model.option
 
 /**
  * 옵션 업데이트 시 결과를 담는 DTO
